liang: implement the heroaircraft shoot function

+++ core/src/com/aircraftwar/android/bullet/AbstractBullet.java
@@ -7,7 +7,7 @@ public class AbstractBullet extends AbstractFlyingObject {
 
     private int power = 10;
 
-    public AbstractBullet(float locationX, float locationY, int speedX, int speedY, int power) {
+    public AbstractBullet(float locationX, float locationY,float speedX, float speedY, int power) {
         super(locationX, locationY, speedX, speedY);
         this.power = power;
     }
