+++ core/src/com/aircraftwar/android/bullet/EnemyBullet.java
@@ -1,7 +1,7 @@
 package com.aircraftwar.android.bullet;
 
 public class EnemyBullet extends AbstractBullet{
-    public EnemyBullet(float locationX, float locationY, int speedX, int speedY, int power) {
+    public EnemyBullet(float locationX, float locationY, float speedX, float speedY, int power) {
         super(locationX, locationY, speedX, speedY, power);
     }
 }
