+++ core/src/com/aircraftwar/android/bullet/HeroBullet.java
@@ -2,7 +2,7 @@ package com.aircraftwar.android.bullet;
 
 public class HeroBullet extends AbstractBullet{
 
-    public HeroBullet(float locationX, float locationY, int speedX, int speedY, int power) {
+    public HeroBullet(float locationX, float locationY, float speedX, float speedY, int power) {
         super(locationX, locationY, speedX, speedY, power);
     }
 }
